% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goal_tables.R
\name{goal_table}
\alias{goal_table}
\title{goal_table}
\usage{
goal_table(year)
}
\arguments{
\item{year}{season}
}
\value{
dataframe
}
\description{
returns dataframe of goals scored and allowed in a given season for each school
}
\details{
requires rvest
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
