% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wncaasocR-package.R
\docType{package}
\name{wncaasocR-package}
\alias{wncaasocR}
\alias{wncaasocR-package}
\title{wncaasocR: Data for Womens NCAA Soccer}
\description{
Scrapes data from http://wosoindependent.com/college/2021/.
}
\author{
\strong{Maintainer}: David Gerth

}
\keyword{internal}
